Create README.md
Project aim:
Implement a web application which is a project that combines front-end, back-end, and AI. It can realize the construction of artificial network without touching the underlying code. 
In the coding of neural networks, especially when high-level frameworks are used, the coupling between different layers is very low. So, we can splice different models like building blocks. we graphically represent the different components in the neural network. We can use drag and drop on the web page to connect different layers of the neural network together. Then send the network to the server. In the back end, we can build the corresponding code and run it on the server. Then return the result to the front-end.